updated backend.template step
I have updated the part where you suggest to use template from s3 URL. The template on the URL has old python version which causes issues when creating the stack as well as the template has some issues which causes bad gateway error on postman down the line. Several people have faced this issue and spent lot of time rectifying it. Updating the template will help the person using this in future.